Add files via upload
This code can be used to extract image files from an insitu video file saved in a TEM. This particular code assumes that the images are saved as a dm4 file. The structure of the folders is second_00, second_01, etc. within a folder named minute_00, minute_01, etc. which is in another folder named hour_00, hour01, etc.
Based on the location of the dm4 files, the images will be chronologically sorted and converted to 8-bit grayscale images. 
The user has to specify the folder location containing the hour_00 (for example) directory. The user also has to specify the output directory where the image sequence needs to be stored.